rename filehandling to resourcehandling
