+++ include/resourcehandling.hpp
