+++ src/main.cpp
@@ -17,7 +17,7 @@
 
 #include <string>
 #include "tui.hpp"
-#include "filehandling.hpp"
+#include "resourcehandling.hpp"
 
 int main() {
     ctui::TUI tui;
