+++ src/resourcehandling.cpp
@@ -3,4 +3,4 @@
 // Full copyright notice in main.cpp
 
 
-#include "filehandling.hpp"
+#include "resourcehandling.hpp"
